hafta_5/AkademikMakaleAsistani_v2/requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# LangChain Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-experimental>=0.0.50

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Integration
requests>=2.31.0
httpx>=0.25.0

# Visualization & Charts
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Streamlit Extensions & UI Components
streamlit-elements>=0.1.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0
streamlit-plotly-events>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced Features
scikit-learn>=1.3.0
networkx>=3.1
pyvis>=0.3.2
transformers>=4.35.0

# System Monitoring (Optional)
psutil>=5.9.0
memory-profiler>=0.60.0

# Additional Utilities
Pillow>=10.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0